fix compilation on macos

+++ src/main.rs
@@ -99,8 +99,8 @@ fn main() {
         let mode = window.current_monitor().unwrap().video_modes().next().unwrap();
         let mon_size = mode.size();
         window.set_outer_position(PhysicalPosition {
-            x: (mon_size.width - win_size.width) / 2,
-            y: (mon_size.height - win_size.height) / 2,
+            x: (mon_size.width as i32 - win_size.width as i32) / 2,
+            y: (mon_size.height as i32 - win_size.height as i32) / 2,
         });
     }
